separate MyProfileImage part from the parent
