+++ lib/models/profile.dart
@@ -1,5 +1,5 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 
 class Profile extends ChangeNotifier {
   final int id;
