+++ lib/screens/my_page/make_profile_page.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:guam_front/commons/back.dart';
 import 'package:guam_front/commons/custom_app_bar.dart';
@@ -7,10 +5,10 @@ import 'package:guam_front/models/profile.dart';
 import 'package:guam_front/providers/stacks/stacks.dart';
 import 'package:guam_front/screens/my_page/profile_filter_value_chip.dart';
 import 'package:hexcolor/hexcolor.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
 
 import '../../providers/user_auth/authenticate.dart';
+import 'my_profile_image.dart';
 
 class MakeProfilePage extends StatefulWidget {
   final Stacks stacksProvider;
@@ -22,7 +20,6 @@ class MakeProfilePage extends StatefulWidget {
 }
 
 class _MakeProfilePageState extends State<MakeProfilePage> {
-  final ImagePicker _picker = ImagePicker();
   final TextEditingController _nicknameController = TextEditingController();
   final TextEditingController _blogController = TextEditingController();
   final TextEditingController _githubIdController = TextEditingController();
@@ -34,11 +31,8 @@ class _MakeProfilePageState extends State<MakeProfilePage> {
     '프론트엔드': <String>[],
     '디자이너': <String>[],
   };
-
   String selectedKey;
   List<String> filterValues;
-
-  PickedFile _imageFile;
   Profile me;
 
   @override
@@ -77,11 +71,6 @@ class _MakeProfilePageState extends State<MakeProfilePage> {
     });
   }
 
-  void takePhoto(ImageSource source) async {
-    final pickedFile = await _picker.getImage(source: source);
-    setState(() => _imageFile = pickedFile);
-  }
-
   @override
   Widget build(BuildContext context) {
     var techStacks = {
@@ -102,7 +91,6 @@ class _MakeProfilePageState extends State<MakeProfilePage> {
 
     Future setProfile(dynamic body) async {
       return await authProvider.setProfile(body).then((successful) {
-        print(successful);
         Navigator.pop(context);
         authProvider.getMyProfile();
       });
@@ -125,124 +113,13 @@ class _MakeProfilePageState extends State<MakeProfilePage> {
                   SizedBox(height: 20),
                   Container(color: Colors.grey),
                   Column(mainAxisAlignment: MainAxisAlignment.start, children: [
-                    imageProfile(context, size),
+                    MyProfileImage(size),
                     _profileInfo(size, techStacks),
                     _authButton(size, setProfile)
                   ]),
                 ],
               ),
             )));
-  }
-
-  Widget imageProfile(BuildContext context, Size size) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Container(
-            padding: EdgeInsets.fromLTRB(25, 20, 0, 0),
-            child: Text("프로필 사진 선택",
-                style: TextStyle(
-                    fontSize: 16,
-                    fontWeight: FontWeight.bold,
-                    color: Colors.black))),
-        Center(
-            child: Stack(children: <Widget>[
-          Container(
-            width: 110,
-            height: 110,
-            child: _imageFile == null
-                ? Icon(Icons.person, color: Colors.white, size: 100)
-                : ClipRRect(
-                    borderRadius: BorderRadius.circular(100),
-                    child: Image(
-                      image: FileImage(File(_imageFile.path)),
-                      fit: BoxFit.fill,
-                    )),
-            decoration: _imageFile == null
-                ? BoxDecoration(boxShadow: [
-                    BoxShadow(
-                        blurRadius: 1,
-                        color: Colors.black.withOpacity(0.5),
-                        offset: Offset(0, 7))
-                  ], shape: BoxShape.circle)
-                : BoxDecoration(),
-          ),
-          Positioned(
-              bottom: 0,
-              right: 0,
-              child: InkWell(
-                  onTap: () {
-                    showModalBottomSheet(
-                        context: context,
-                        builder: ((builder) => bottomSheet(size)));
-                  },
-                  child: Container(
-                    decoration: BoxDecoration(
-                      color: Colors.white,
-                      shape: BoxShape.circle,
-                      border: Border.all(
-                        color: Colors.white,
-                        width: 1,
-                      ),
-                      gradient: LinearGradient(
-                          colors: [
-                            HexColor("4F34F3"),
-                            HexColor("3EF7FF"),
-                          ],
-                          begin: FractionalOffset(1.0, 0.0),
-                          end: FractionalOffset(0.0, 0.0),
-                          stops: [0, 1],
-                          tileMode: TileMode.clamp),
-                    ),
-                    child: Icon(
-                      Icons.add,
-                      color: Colors.white,
-                      size: 30,
-                    ),
-                  )))
-        ])),
-      ],
-    );
-  }
-
-  Widget bottomSheet(Size size) {
-    return Container(
-      height: 130,
-      width: size.width,
-      margin: EdgeInsets.symmetric(
-        horizontal: 20,
-        vertical: 20,
-      ),
-      child: Column(
-        children: <Widget>[
-          Text(
-            "프로필 사진을 설정해주세요.",
-            style: TextStyle(
-              fontSize: 20,
-            ),
-          ),
-          SizedBox(
-            height: 25,
-          ),
-          Row(mainAxisAlignment: MainAxisAlignment.center, children: <Widget>[
-            TextButton.icon(
-              icon: Icon(Icons.camera),
-              onPressed: () {
-                takePhoto(ImageSource.camera);
-              },
-              label: Text("카메라"),
-            ),
-            TextButton.icon(
-              icon: Icon(Icons.image),
-              onPressed: () {
-                takePhoto(ImageSource.gallery);
-              },
-              label: Text("갤러리"),
-            ),
-          ])
-        ],
-      ),
-    );
   }
 
   Widget _profileInfo(Size size, Map<dynamic, List<dynamic>> techStacks) {
