+++ lib/screens/my_page/my_page.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:guam_front/providers/stacks/stacks.dart';
-import '../user_auth/kakao_login.dart';
-import '../../commons/custom_app_bar.dart';
-import 'make_profile_page.dart';
 import 'package:provider/provider.dart';
+
+import '../../commons/custom_app_bar.dart';
 import '../../providers/user_auth/authenticate.dart';
+import '../user_auth/kakao_login.dart';
+import 'make_profile_page.dart';
 import 'my_profile.dart';
 
 class MyPage extends StatelessWidget {
@@ -16,12 +17,19 @@ class MyPage extends StatelessWidget {
     return Scaffold(
       appBar: CustomAppBar(
         title: "프로필",
-        trailing: IconButton(
-          icon: Icon(
-            Icons.settings,
-            color: Colors.black,
-          ),
-        ),
+        trailing: authProvider.profileExists()
+            ? TextButton(
+                child: Text(
+                  'EDIT',
+                  style: TextStyle(
+                      color: Colors.black,
+                      fontSize: 14,
+                      fontWeight: FontWeight.bold),
+                ),
+                onPressed: () => Navigator.of(context).push(MaterialPageRoute(
+                    builder: (context) => MakeProfilePage(stacksProvider))),
+              )
+            : null,
       ),
       body: authProvider.loading
           ? Center(child: CircularProgressIndicator())
